Crate daggy

daggy is a directed acyclic graph data structure library.

The most prominent type is Dag - a wrapper around petgraph (http://bluss.github.io/petulant-avenger-graphlibrary/doc/petgraph/index.html)'s [Graph] (http://bluss.github.io/petulant-avenger-graphlibrary/doc/petgraph/graph/struct.Graph.html) data structure, exposing a refined API targeted towards directed acyclic graph related functionality.

The Walker trait defines a variety of useful methods for traversing any graph type. Its methods behave similarly to iterator types, however Walkers do not require borrowing the graph. This means that we can still safely mutably borrow from the graph whilst we traverse it.

Re-exports

pub extern crate petgraph;

pub use walker::Walker;

Modules

walker
Walker is a trait providing a variety of useful methods for traversing graph types.

Structs

Children
A Walker type that can be used to step through the children of some parent node.

Dag
A Directed acyclic graph (DAG) data structure.

EdgeIndex
Edge identifier.

EdgeIndices
An iterator yielding multiple EdgeIndexs, returned by the Graph::add_edges method.

EdgeWeightsMut
Iterator yielding mutable access to all edge weights.

NodeIndex
Node identifier.

NodeWeightsMut
Iterator yielding mutable access to all node weights.

Parents
A Walker type that can be used to step through the parents of some child node.

WouldCycle
An error returned by the Dag::add_edge method in the case that adding an edge would have caused the graph to cycle.

Type Definitions

PetGraph
The Petgraph to be used internally within the Dag for storing/managing nodes and edges.

RawEdges
Read only access into a Dag's internal edge array.

RawNodes
Read only access into a Dag's internal node array.

RecursiveWalk
An alias to simplify the Recursive Walker type returned by Dag.