Function daggy::petgraph::algo::scc

pub fn scc<N, E, Ty, Ix>(g: &Graph<N, E, Ty, Ix>) -> Vec<Vec<NodeIndex<Ix>>> where Ix: IndexType, Ty: EdgeType

Compute the strongly connected components using Kosaraju's algorithm.

Return a vector where each element is an scc.

For an undirected graph, the sccs are simply the connected components.