Module daggy::petgraph::graph

Graph<N, E, Ty, Ix> is a graph datastructure using an adjacency list representation.

Structs

Edge
The graph's edge type.

EdgeIndex
Edge identifier.

EdgeWeightsMut
Iterator yielding mutable access to all edge weights.

Edges
Iterator over the edges of a node.

Graph
Graph<N, E, Ty, Ix> is a graph datastructure using an adjacency list representation.

Neighbors
Iterator over the neighbors of a node.

Node
The graph's node type.

NodeIndex
Node identifier.

NodeWeightsMut
Iterator yielding mutable access to all node weights.

WalkEdges
A “walker” object that can be used to step through the edge list of a node.

WithoutEdges
An iterator over either the nodes without edges to them or from them.

Traits

GraphIndex
A GraphIndex is a node or edge index.

IndexType
Trait for the unsigned integer type used for node and edge indices.

Type Definitions

DefIndex