Struct daggy::petgraph::graph::WalkEdges

pub struct WalkEdges<Ix> where Ix: IndexType {
    // some fields omitted
}

A “walker” object that can be used to step through the edge list of a node.

See .walk_edges_directed() for more information.

Methods

impl<Ix> WalkEdges<Ix> where Ix: IndexType

fn next<N, E, Ty>(&mut self, g: &Graph<N, E, Ty, Ix>) -> Option<EdgeIndex<Ix>> where Ty: EdgeType

Fetch the next edge index in the walk for graph g.

fn next_neighbor<N, E, Ty>(&mut self, g: &Graph<N, E, Ty, Ix>) -> Option<(EdgeIndex<Ix>, NodeIndex<Ix>)> where Ty: EdgeType

Fetch the next edge index and the next node index in the walk for graph g.

The next node indices are always the others than the starting point where the WalkEdges value was created. For an Outgoing walk, the target nodes, for an Incoming walk, the source nodes of the edge.