Struct daggy::petgraph::Dfs

pub struct Dfs<N, VM> {
    pub stack: Vec<N>,
    pub discovered: VM,
}

A depth first search (DFS) of a graph.

Using a Dfs you can run a traversal over a graph while still retaining mutable access to it, if you use it like the following example:

use petgraph::{Graph, Dfs};

let mut graph = Graph::<_,()>::new();
let a = graph.add_node(0);

let mut dfs = Dfs::new(&graph, a);
while let Some(nx) = dfs.next(&graph) {
    // we can access `graph` mutably here still
    graph[nx] += 1;
}

assert_eq!(graph[a], 1);

Note: The algorithm may not behave correctly if nodes are removed during iteration. It may not necessarily visit added nodes or edges.

Fields

stack

The stack of nodes to visit

discovered

The map of discovered nodes