Struct daggy::WalkChildren

pub struct WalkChildren<Ix: IndexType> {
    // some fields omitted
}

A "walker" object that can be used to step through the children of some parent node.

Methods

impl<Ix> WalkChildren<Ix> where Ix: IndexType

fn next<N, E>(&mut self, dag: &Dag<N, E, Ix>) -> Option<EdgeIndex<Ix>>

Fetch the next child edge index in the walk for the given Dag.

fn next_child<N, E>(&mut self, dag: &Dag<N, E, Ix>) -> Option<(EdgeIndex<Ix>, NodeIndex<Ix>)>

Fetch the EdgeIndex and NodeIndex to the next child in the walk for the given Dag.