Module daggy::walker

Walker is a trait providing a variety of useful methods for traversing graph types.

Structs

Chain

Walks the entirety of a before walking the entirety of b.

Cycle

A walker that repeats its internal walker endlessly.

Filter

A walker that applies some given predicate to each element returned by its walker.

Inspect

A walker that calls a function with a reference to each index pair before yielding them.

Peekable

A walker that has a .peek(&graph) method that returns an optional next neighbor.

Recursive

Recursively walks a graph using the recursive function recursive_fn.

Skip

A walker that skips the first n steps of this walk, and then yields all further steps.

SkipWhile

A walker that invokes the predicate on elements until it returns false.

Take

A walker that yields the first n steps of this walk.

TakeWhile

A walker that yields elements so long as the predicate returns true.