Struct daggy::walker::Filter

pub struct Filter<W, P> {
    // some fields omitted
}

A walker that applies some given predicate to each element returned by its walker. The only index pairs that will be yielded are those that make the predicate evaluate to true.