Struct daggy::walker::Peekable

pub struct Peekable<G, Ix, W> where Ix: IndexType { /* fields omitted */ }

A walker that has a .peek(&graph) method that returns an optional next neighbor.

Methods

impl<G, Ix, W> Peekable<G, Ix, W> where Ix: IndexType, W: Walker<G, Index=Ix>

fn peek(&mut self, graph: &G) -> Option<IndexPair<Ix>>

The edge node index pair of the neighbor at the next step in our walk of the given graph.

fn peek_edge(&mut self, graph: &G) -> Option<EdgeIndex<Ix>>

The edge index of the neighbor at the next step in our walk of the given graph.

fn peek_node(&mut self, graph: &G) -> Option<NodeIndex<Ix>>

The node index of the neighbor at the next step in our walk of the given graph.