Struct daggy::walker::Recursive

pub struct Recursive<G, Ix, F> { /* fields omitted */ }

Recursively walks a graph using the recursive function recursive_fn.

Methods

impl<G, Ix, F> Recursive<G, Ix, F>

pub fn new(start: NodeIndex<Ix>, recursive_fn: F) -> Self where
Ix: IndexType,
F: FnMut(&G, NodeIndex<Ix>) -> Option<IndexPair<Ix>>,

Construct a new Recursive Walker starting from the node at the given index.