Module daggy::petgraph::graph

Graph<N, E, Ty, Ix> is a graph datastructure using an adjacency list representation.

Structs

Edge
The graph's edge type.

EdgeIndex
Edge identifier.

EdgeIndices
Iterator over the edge indices of a graph.

EdgeWeightsMut
Iterator yielding mutable access to all edge weights.

Edges
Iterator over the edges of a node.

Externals
An iterator over either the nodes without edges to them or from them.

Graph
Graph<N, E, Ty, Ix> is a graph datastructure using an adjacency list representation.

Neighbors
Iterator over the neighbors of a node.

Node
The graph's node type.

NodeIndex
Node identifier.

NodeIndices
Iterator over the node indices of a graph.

NodeWeightsMut
Iterator yielding mutable access to all node weights.

WalkEdges
Deprecated.

WalkNeighbors
A “walker” object that can be used to step through the edge list of a node.

Traits

GraphIndex
A GraphIndex is a node or edge index.

IndexType
Trait for the unsigned integer type used for node and edge indices.

Functions

edge_index
Short version of EdgeIndex::new

node_index
Short version of NodeIndex::new

Type Definitions

DefIndex